sprite_similarity/requirements.txt
# ipyplot==1.1.1
# ipython==8.9.0
matplotlib==3.5.1
numpy==1.23.3
pandas==1.3.4
Pillow==9.4.0
scikit_image==0.19.1
requests==2.24.0
